# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRM для Telegram ботов API",
        "description": "API документация для CRM системы управления Telegram ботами",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API V1"
        }
    ],
    "paths": {
        "/folders": {
            "get": {
                "tags": [
                    "Folders"
                ],
                "summary": "Получить список папок",
                "description": "Возвращает список папок с возможностью фильтрации по родительской папке и пагинации",
                "operationId": "359b9530f93593f7ad4763a8feeb7782",
                "parameters": [
                    {
                        "name": "parent_id",
                        "in": "query",
                        "description": "ID родительской папки (null для корневых папок)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "nullable": true
                        }
                    },
                    {
                        "name": "paginate",
                        "in": "query",
                        "description": "Количество элементов на странице (0 - без пагинации)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Folder"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Folders"
                ],
                "summary": "Создать новую папку",
                "description": "Создаёт новую папку в медиа-менеджере",
                "operationId": "a17f6776f0d1c0a3965276e14b0ec7ff",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Название папки",
                                        "type": "string",
                                        "example": "Новая папка"
                                    },
                                    "parent_id": {
                                        "description": "ID родительской папки",
                                        "type": "integer",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "slug": {
                                        "description": "URL-slug (генерируется автоматически)",
                                        "type": "string",
                                        "example": "novaya-papka",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Папка успешно создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Folder"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/folders/{id}": {
            "get": {
                "tags": [
                    "Folders"
                ],
                "summary": "Получить информацию о папке",
                "description": "Возвращает детальную информацию о конкретной папке",
                "operationId": "25b407039d70dadd2e2b8e3968914d10",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID папки",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Folder"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Папка не найдена"
                    }
                }
            },
            "put": {
                "tags": [
                    "Folders"
                ],
                "summary": "Обновить папку",
                "description": "Обновляет данные папки",
                "operationId": "40be455f942c06ac56a40e59e5b8d89a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID папки",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Обновлённая папка"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "position": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Папка успешно обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Folder"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Папка не найдена"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Folders"
                ],
                "summary": "Удалить папку",
                "description": "Удаляет папку и все её содержимое (файлы и вложенные папки)",
                "operationId": "e6eb7f200bbccb95c9eb49ca71c2e27b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID папки",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Папка успешно удалена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Папка успешно удалена"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Папка не найдена"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                }
            }
        },
...
        "/folders/update-positions": {
            "post": {
                "tags": [
                    "Folders"
                ],
                "summary": "Обновить позиции папок",
                "description": "Массовое обновление позиций папок для реализации drag & drop",
                "operationId": "bb7372ea45116ab7647ac622131a24bb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "folders"
                                ],
                                "properties": {
                                    "folders": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "position": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Позиции успешно обновлены",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Позиции папок успешно обновлены"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Folder": {
                "title": "Folder",
                "description": "Модель папки медиа-менеджера",
                "required": [
                    "id",
                    "name",
                    "slug"
                ],
                "properties": {
                    "id": {
                        "description": "ID папки",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Название папки",
                        "type": "string",
                        "example": "Общая"
                    },
                    "slug": {
                        "description": "URL-slug папки",
                        "type": "string",
                        "example": "common"
                    },
                    "src": {
                        "description": "Иконка папки",
                        "type": "string",
                        "example": "folder"
                    },
                    "parent_id": {
                        "description": "ID родительской папки",
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "position": {
                        "description": "Позиция для сортировки",
                        "type": "integer",
                        "example": 0
                    },
                    "count": {
                        "description": "Количество файлов в папке",
                        "type": "integer",
                        "example": 5
                    },
                    "countFolder": {
                        "description": "Количество вложенных папок",
                        "type": "integer",
                        "example": 2
                    },
                    "parent": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Folder"
                            }
                        ],
                        "nullable": true,
                        "description": "Родительская папка"
                    },
                    "children": {
                        "description": "Вложенные папки",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Folder"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Введите токен в формате: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}